rename fuzz_cfg_load.py to cfg_load.py
